refactor(md, term, cli): Move ANSI stream machinery into `jp_term`
`ansi` and `shade` move from `jp_md` to `jp_term`, along with the
region-background types they interpret, which land in a new
`jp_term::background`. Nothing about what they do is markdown: `segments`
tokenizes an escape stream, `AnsiState` tracks what a stream left active,
and `ShadedWriter` keeps a background showing across cursor rewrites and
split escapes. `jp_md` gains a `jp_term` dependency and imports them like
any other consumer.

The move puts them where the printer can reach them. `jp_printer` depends
on `jp_term` and deliberately not on `jp_md`, so shading a prompt meant
either pulling a markdown renderer into the output layer or writing the
state machine a second time.

`DefaultBackground`, `BackgroundFill`, and `line_fill` leave
`jp_md::format` for `jp_term::background`, so the types and the one place
that interprets them travel together. `AnsiState`'s methods widen from
`pub(crate)` to `pub` now that its callers are in another crate.
