+++ crates/jp_md/src/format.rs
@@ -1,17 +1,20 @@
 //! Markdown formatting utilities.
 
-use std::{borrow::Cow, fmt, sync::LazyLock};
+use std::{fmt, sync::LazyLock};
 
 use comrak::{
     Arena,
     nodes::{NodeList, NodeValue},
     options::{Extension, ListStyleType, Render},
 };
+use jp_term::{
+    ansi::{self, AnsiState, Segment},
+    background::{DefaultBackground, line_fill},
+};
 use syntect::{highlighting::Theme, parsing::SyntaxSet};
 use two_face::syntax;
 
 use crate::{
-    ansi::{self, AnsiState, Segment},
     render::{self, HrOptions, RenderOptions},
     table::TableOptions,
     theme,
@@ -33,19 +36,6 @@ const DEFAULT_WIDTH: usize = 80;
 /// Default maximum column width for tables.
 const DEFAULT_TABLE_MAX_COL_WIDTH: usize = 40;
 
-/// How a default background color fills each line.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum BackgroundFill {
-    /// Fill to the last visible character on the line.
-    Content,
-
-    /// Fill to a fixed column width (padding with spaces if needed).
-    Column(usize),
-
-    /// Fill to the end of the terminal window via `\x1b[K`.
-    Terminal,
-}
-
 /// Controls how horizontal rules (`---`) are rendered in terminal output.
 #[derive(Debug, Clone, Copy, Default)]
 pub enum HrStyle {
@@ -56,17 +46,6 @@ pub enum HrStyle {
     /// line width (based on `wrap_width`).
     #[default]
     Line,
-}
-
-/// A default background color applied to all content, with a fill mode
-/// controlling how far it extends on each line.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DefaultBackground {
-    /// SGR background parameter, e.g. `"48;5;236"` or `"48;2;80;73;69"`.
-    pub param: String,
-
-    /// How far the background extends on each line.
-    pub fill: BackgroundFill,
 }
 
 /// Per-call options for [`Formatter::format_terminal_with`].
@@ -482,28 +461,6 @@ fn ends_with_tight_list<'a>(root: &'a comrak::nodes::AstNode<'a>) -> bool {
         last.data().value,
         NodeValue::List(NodeList { tight: true, .. })
     )
-}
-
-/// The text that extends a background from `column` to the end of the line.
-///
-/// The one place [`BackgroundFill`] is interpreted.
-/// Every writer that maintains a region background consults this so the three
-/// modes cannot drift apart: `Content` adds nothing, `Terminal` defers to the
-/// terminal's erase-to-end-of- line, and `Column` pads with real spaces — the
-/// only form a host that lays out its own sub-window (an `fzf` preview pane)
-/// renders, since it does not implement the erase.
-///
-/// The caller is responsible for having the background active before writing
-/// the result.
-pub(crate) fn line_fill(fill: BackgroundFill, column: usize) -> Cow<'static, str> {
-    match fill {
-        BackgroundFill::Content => Cow::Borrowed(""),
-        BackgroundFill::Terminal => Cow::Borrowed("\x1b[K"),
-        BackgroundFill::Column(target) => match target.saturating_sub(column) {
-            0 => Cow::Borrowed(""),
-            pad => Cow::Owned(" ".repeat(pad)),
-        },
-    }
 }
 
 /// Render an inter-block separator (blank line) with optional background fill.
