+++ crates/jp_term/src/lib.rs
@@ -1,5 +1,8 @@
 #![allow(clippy::print_stdout, clippy::print_stderr)]
 
+pub mod ansi;
+pub mod background;
 pub mod osc;
+pub mod shade;
 pub mod table;
 pub mod width;
